fix status_data.py circular import error
- decouple StatusData class into another file, this was causing circular
  import.
- update test for skelet
- move get_now_str() function into date.py file
- copy changes in opl dir into core/opl dir
